Backend/requirements.txt
# Django core (Django 4.2 LTS for Python 3.9 compatibility)
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# API documentation
drf-spectacular>=0.27,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication (JWT structure ready for real auth)
djangorestframework-simplejwt>=5.3,<6.0

# AI Integration
google-generativeai>=0.5,<1.0

# Environment variables
python-dotenv>=1.0,<2.0

# Development
django-debug-toolbar>=4.2,<5.0

# File handling
Pillow>=9.0,<11.0
